🚚 Rename `SearchWrapper` layout component
